moved the files to proper folder
